bpf:cgroups: do advanced cgroups tracking only if deployment mode is set
If Tetragon deployment mode is known perform advanced cgroups tracking,
otherwise return as early as possible when probing cgroups
configuration.

+++ pkg/sensors/config/confmap/confmap.go
@@ -21,7 +21,7 @@ type TetragonConfKey struct {
 }
 
 type TetragonConfValue struct {
-	Mode            uint32 `align:"mode"`               // Deployment mode
+	Mode            int32  `align:"mode"`               // Deployment mode
 	LogLevel        uint32 `align:"loglevel"`           // Tetragon log level
 	PID             uint32 `align:"pid"`                // Tetragon PID for debugging purpose
 	NSPID           uint32 `align:"nspid"`              // Tetragon PID in namespace for debugging purpose
@@ -115,7 +115,7 @@ func UpdateTgRuntimeConf(mapDir string, nspid int) error {
 
 	k := &TetragonConfKey{Key: 0}
 	v := &TetragonConfValue{
-		Mode:            deployMode,
+		Mode:            int32(deployMode),
 		LogLevel:        uint32(logger.GetLogLevel()),
 		TgCgrpHierarchy: cgroups.GetCgrpHierarchyID(),
 		TgCgrpSubsysIdx: cgroups.GetCgrpSubsystemIdx(),
